+++ scripts/spammys_chat_filter/SpammyChatFilter.cs
@@ -6,27 +6,26 @@ using System.Collections.Generic;
 public class SpammyChatFilter : IHoldfastSharedMethods {
     private InputField f1MenuInputField;
     private int damage = 5;
-    private int mute_threshold = 3; 
+    private int muteThreshold = 3;
     private string reason = "Chat Filter: Please refrain from saying banned words in chat.";
-    private string mute_message = "Chat Filter: You have now been temp muted from text chat.";
-    private List<string> banned_words = new List<string>()
-    {
+    //The three stars will capture all the currently censored words in-game (e.g. 'nazi'). These are censored before the message is sent, so the actual words themselves cannot be checked.
+    private List<string> bannedWordsList = new List<string>() {
         "***"
-    }; //The three stars will capture all the currently censored words in-game (for example, 'nazi' etc.). These are censored before the message is sent, so the actual words themselves cannot be slapped.
+    };
     private Dictionary<int, int> playerSlapList = new Dictionary<int, int>(); //Maps playerId and the times the player has been slapped
 
     public void OnIsServer(bool server) {
+        Debug.Log("SCF: Starting load...");
         //Get all the canvas items in the game
         var canvases = Resources.FindObjectsOfTypeAll<Canvas>();
         for (int i = 0; i < canvases.Length; i++) {
-            //Find the one that's called "Game Console Panel"
             if (string.Compare(canvases[i].name, "Game Console Panel", true) == 0) {
                 //Inside this, now we need to find the input field where the player types messages.
                 f1MenuInputField = canvases[i].GetComponentInChildren<InputField>(true);
                 if (f1MenuInputField != null) {
-                    Debug.Log("Found the Game Console Panel");
+                    Debug.Log("SCF: Found the Game Console Panel");
                 } else {
-                    Debug.Log("Game Console Panel not found");
+                    Debug.Log("SCF: Game Console Panel not found! This may cause stability issues.");
                 }
                 break;
             }
@@ -34,27 +33,13 @@ public class SpammyChatFilter : IHoldfastSharedMethods {
     }
 
     public void OnTextMessage(int playerId, TextChatChannel channel, string text) {
-        // First convert it to lower case, this prevents players circumventing it with odd casing
         string uCaseText = text.ToLower();
-        foreach(string banned_word in banned_words) {
-            if (uCaseText.Contains(banned_word)) {
+        foreach (string bannedWord in bannedWordsList) {
+            if (uCaseText.Contains(bannedWord)) {
+                Debug.Log(string.Format("SCF: Caught banned word {0} from player {1}", bannedWord, playerId));
                 if (f1MenuInputField != null) {
-                    var rcSlapCommand = string.Format("serverAdmin slap {0} {1} {2}", playerId, damage, reason);
-                    f1MenuInputField.onEndEdit.Invoke(rcSlapCommand);
-                    var rcWarnCommand = string.Format("serverAdmin privateMessage {0} {1}", playerId, reason);
-                    f1MenuInputField.onEndEdit.Invoke(rcWarnCommand);
-
-                    //Increment the player's slap count
-                    int slapCount;
-                    playerSlapList.TryGetValue(playerId, out slapCount);
-                    playerSlapList[playerId] = slapCount + 1;
-                    // Check if the player needs to be muted
-                    if (playerSlapList[playerId] == mute_threshold) {
-                        // Temp mute the player - this only needs to happen once per round
-                        var rcMuteCommand = string.Format("serverAdmin chatMute {0} {1}", playerId, "You are now auto-temp muted.");
-                        f1MenuInputField.onEndEdit.Invoke(rcMuteCommand);
-                    }
-                    break; // Break so the entire list is not parsed every time - don't punish players for more than one word
+                    slapAndMessagePlayer(playerId);
+                    checkToMutePlayer(playerId);
                 }
             }
         }
@@ -67,26 +52,24 @@ public class SpammyChatFilter : IHoldfastSharedMethods {
                 continue;
             }
 
-            //so first variable should be the mod id
-            if (splitData[0] == "2549637575") {
-                //the second variable should be the variable type
-                if (splitData[1] == "schat_filter_slap_damage") {
-                    //and the third variable should be the variable value
+            if (splitData[0] == "scf") {
+                if (splitData[1] == "scf_slap_damage") {
                     if (!int.TryParse(splitData[2], out damage)) {
-                        Debug.Log("Tried parsing slap_damage but invalid format was found.");
+                        Debug.Log("SCF: Tried parsing slap_damage but invalid format was found.");
+                    } else {
+                        Debug.Log("SCF: Parsed custom parameter scf_slap_damage");
                     }
-                }
-                //similarly, reason is the variable type
-                else if (splitData[1] == "schat_filter_slap_reason") {
-                    //fill the reason using the variable value
+                } else if (splitData[1] == "scf_slap_reason") {
                     reason = splitData[2];
-                } else if (splitData[1] == "schat_filter_banned_word") {
-                    // Add a banned word to the list
-                    banned_words.Add(splitData[2]);
-                } else if (splitData[1] == "schat_filter_mute_limit") {
-                    // Change how many slaps before user is muted
-                    if (!int.TryParse(splitData[2], out mute_threshold)) {
-                        Debug.Log("Tried parsing mute_threshold but invalid format was found.");
+                    Debug.Log("SCF: Parsed custom parameter scf_slap_reason");
+                } else if (splitData[1] == "scf_banned_word") {
+                    bannedWordsList.Add(splitData[2]);
+                    Debug.Log(string.Format("SCF: Added banned word {0}", splitData[2]));
+                } else if (splitData[1] == "scf_mute_threshold") {
+                    if (!int.TryParse(splitData[2], out muteThreshold)) {
+                        Debug.Log("SCF: Tried parsing mute_threshold but invalid format was found.");
+                    } else {
+                        Debug.Log("SCF: Parsed custom parameter scf_mute_threshold");
                     }
                 }
             }
@@ -95,6 +78,7 @@ public class SpammyChatFilter : IHoldfastSharedMethods {
 
     public void OnPlayerJoined(int playerId, ulong steamId, string playerName, string regimentTag, bool isBot) {
         // Add the player to the dictionary
+        Debug.Log(string.Format("SCF: Player {0} ({1}, steamId = {2}) joined the server, player added to watch list", playerId, playerName, steamId));
         playerSlapList.Add(playerId, 0);
     }
 
@@ -102,6 +86,30 @@ public class SpammyChatFilter : IHoldfastSharedMethods {
         // If a player has left, remove them from the dictionary
         if (playerSlapList.ContainsKey(playerId)) {
             playerSlapList.Remove(playerId);
+            Debug.Log("SCF: Player {0} left the server, player removed from watch list");
+        }
+    }
+
+    public void slapAndMessagePlayer(int playerId) {
+        var rcSlapCommand = string.Format("serverAdmin slap {0} {1} {2}", playerId, damage, reason);
+        Debug.Log(string.Format("SCF: Slapping and warning player {0}", playerId));
+        f1MenuInputField.onEndEdit.Invoke(rcSlapCommand);
+        var rcWarnCommand = string.Format("serverAdmin privateMessage {0} {1}", playerId, reason);
+        f1MenuInputField.onEndEdit.Invoke(rcWarnCommand);
+    }
+
+    public void checkToMutePlayer(int playerId) {
+        Debug.Log(string.Format("SCF: Checking player {0}'s banned word usage...", playerId));
+        playerSlapList.TryGetValue(playerId, out var slapCount);
+        playerSlapList[playerId] = slapCount + 1;
+        Debug.Log(string.Format("SCF: Player {0} has been warned {1} times", playerId, slapCount));
+        // Check if the player needs to be muted
+        if (playerSlapList[playerId] == muteThreshold) {
+            Debug.Log(string.Format("SCF: Player {0} has exceeded the mute threshold, muting player...", playerId));
+            var rcMuteCommand = string.Format("serverAdmin chatMute {0}", playerId);
+            f1MenuInputField.onEndEdit.Invoke(rcMuteCommand);
+        } else if (playerSlapList[playerId] > muteThreshold) {
+            Debug.Log(string.Format("SCF: Player {0} has already been chat muted", playerId));
         }
     }
 
@@ -225,9 +233,9 @@ public class SpammyChatFilter : IHoldfastSharedMethods {
     public void OnVehiclePacket(int vehicleId, Vector2 inputAxis, bool shift, bool strafe, PlayerVehicleActions[] actionCollection) {
     }
 
-    public void OnOfficerOrderStart(int officerPlayerId, OfficerOrderType officerOrderType, Vector3 orderPosition, float orderRotationY, int voicePhraseRandomIndex) {
+    public void OnOfficerOrderStart(int officerPlayerId, HighCommandOrderType highCommandOrderType, Vector3 orderPosition, float orderRotationY, int voicePhraseRandomIndex) {
     }
 
-    public void OnOfficerOrderStop(int officerPlayerId, OfficerOrderType officerOrderType) {
+    public void OnOfficerOrderStop(int officerPlayerId, HighCommandOrderType highCommandOrderType) {
     }
 }